Extract out ModelStub
